+++ sdk/resourcemanager/dependencymap/armdependencymap/discoverysources_client.go
@@ -42,7 +42,7 @@ func NewDiscoverySourcesClient(subscriptionID string, credential azcore.TokenCre
 // BeginCreateOrUpdate - Create a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - sourceName - discovery source resource
@@ -69,7 +69,7 @@ func (client *DiscoverySourcesClient) BeginCreateOrUpdate(ctx context.Context, r
 // CreateOrUpdate - Create a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *DiscoverySourcesClient) createOrUpdate(ctx context.Context, resourceGroupName string, mapName string, sourceName string, resource DiscoverySourceResource, options *DiscoverySourcesClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DiscoverySourcesClient.BeginCreateOrUpdate"
@@ -115,7 +115,7 @@ func (client *DiscoverySourcesClient) createOrUpdateCreateRequest(ctx context.Co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -128,7 +128,7 @@ func (client *DiscoverySourcesClient) createOrUpdateCreateRequest(ctx context.Co
 // BeginDelete - Delete a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - sourceName - discovery source resource
@@ -154,7 +154,7 @@ func (client *DiscoverySourcesClient) BeginDelete(ctx context.Context, resourceG
 // Delete - Delete a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *DiscoverySourcesClient) deleteOperation(ctx context.Context, resourceGroupName string, mapName string, sourceName string, options *DiscoverySourcesClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DiscoverySourcesClient.BeginDelete"
@@ -200,16 +200,15 @@ func (client *DiscoverySourcesClient) deleteCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
 // Get - Get a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - sourceName - discovery source resource
@@ -260,7 +259,7 @@ func (client *DiscoverySourcesClient) getCreateRequest(ctx context.Context, reso
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -277,7 +276,7 @@ func (client *DiscoverySourcesClient) getHandleResponse(resp *http.Response) (Di
 
 // NewListByMapsResourcePager - List DiscoverySourceResource resources by MapsResource
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - options - DiscoverySourcesClientListByMapsResourceOptions contains the optional parameters for the DiscoverySourcesClient.NewListByMapsResourcePager
@@ -325,7 +324,7 @@ func (client *DiscoverySourcesClient) listByMapsResourceCreateRequest(ctx contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -343,7 +342,7 @@ func (client *DiscoverySourcesClient) listByMapsResourceHandleResponse(resp *htt
 // BeginUpdate - Update a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - sourceName - discovery source resource
@@ -370,7 +369,7 @@ func (client *DiscoverySourcesClient) BeginUpdate(ctx context.Context, resourceG
 // Update - Update a DiscoverySourceResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *DiscoverySourcesClient) update(ctx context.Context, resourceGroupName string, mapName string, sourceName string, properties DiscoverySourceResourceTagsUpdate, options *DiscoverySourcesClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DiscoverySourcesClient.BeginUpdate"
@@ -416,7 +415,7 @@ func (client *DiscoverySourcesClient) updateCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}