+++ sdk/resourcemanager/dependencymap/armdependencymap/maps_client.go
@@ -42,7 +42,7 @@ func NewMapsClient(subscriptionID string, credential azcore.TokenCredential, opt
 // BeginCreateOrUpdate - Create a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - resource - Resource create parameters.
@@ -68,7 +68,7 @@ func (client *MapsClient) BeginCreateOrUpdate(ctx context.Context, resourceGroup
 // CreateOrUpdate - Create a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) createOrUpdate(ctx context.Context, resourceGroupName string, mapName string, resource MapsResource, options *MapsClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginCreateOrUpdate"
@@ -110,7 +110,7 @@ func (client *MapsClient) createOrUpdateCreateRequest(ctx context.Context, resou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -123,7 +123,7 @@ func (client *MapsClient) createOrUpdateCreateRequest(ctx context.Context, resou
 // BeginDelete - Delete a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - options - MapsClientBeginDeleteOptions contains the optional parameters for the MapsClient.BeginDelete method.
@@ -147,7 +147,7 @@ func (client *MapsClient) BeginDelete(ctx context.Context, resourceGroupName str
 // Delete - Delete a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) deleteOperation(ctx context.Context, resourceGroupName string, mapName string, options *MapsClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginDelete"
@@ -189,16 +189,15 @@ func (client *MapsClient) deleteCreateRequest(ctx context.Context, resourceGroup
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
 // BeginExportDependencies - Export dependencies
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - body - The content of the action request
@@ -224,7 +223,7 @@ func (client *MapsClient) BeginExportDependencies(ctx context.Context, resourceG
 // ExportDependencies - Export dependencies
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) exportDependencies(ctx context.Context, resourceGroupName string, mapName string, body ExportDependenciesRequest, options *MapsClientBeginExportDependenciesOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginExportDependencies"
@@ -239,7 +238,7 @@ func (client *MapsClient) exportDependencies(ctx context.Context, resourceGroupN
 	if err != nil {
 		return nil, err
 	}
-	if !runtime.HasStatusCode(httpResp, http.StatusAccepted) {
+	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusAccepted) {
 		err = runtime.NewResponseError(httpResp)
 		return nil, err
 	}
@@ -266,7 +265,7 @@ func (client *MapsClient) exportDependenciesCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -279,7 +278,7 @@ func (client *MapsClient) exportDependenciesCreateRequest(ctx context.Context, r
 // Get - Get a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - options - MapsClientGetOptions contains the optional parameters for the MapsClient.Get method.
@@ -325,7 +324,7 @@ func (client *MapsClient) getCreateRequest(ctx context.Context, resourceGroupNam
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -343,7 +342,7 @@ func (client *MapsClient) getHandleResponse(resp *http.Response) (MapsClientGetR
 // BeginGetConnectionsForProcessOnFocusedMachine - Get network connections of a process
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - body - The content of the action request
@@ -369,7 +368,7 @@ func (client *MapsClient) BeginGetConnectionsForProcessOnFocusedMachine(ctx cont
 // GetConnectionsForProcessOnFocusedMachine - Get network connections of a process
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) getConnectionsForProcessOnFocusedMachine(ctx context.Context, resourceGroupName string, mapName string, body GetConnectionsForProcessOnFocusedMachineRequest, options *MapsClientBeginGetConnectionsForProcessOnFocusedMachineOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginGetConnectionsForProcessOnFocusedMachine"
@@ -411,9 +410,8 @@ func (client *MapsClient) getConnectionsForProcessOnFocusedMachineCreateRequest(
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
 		return nil, err
@@ -424,7 +422,7 @@ func (client *MapsClient) getConnectionsForProcessOnFocusedMachineCreateRequest(
 // BeginGetConnectionsWithConnectedMachineForFocusedMachine - Get network connections between machines
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - body - The content of the action request
@@ -450,7 +448,7 @@ func (client *MapsClient) BeginGetConnectionsWithConnectedMachineForFocusedMachi
 // GetConnectionsWithConnectedMachineForFocusedMachine - Get network connections between machines
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) getConnectionsWithConnectedMachineForFocusedMachine(ctx context.Context, resourceGroupName string, mapName string, body GetConnectionsWithConnectedMachineForFocusedMachineRequest, options *MapsClientBeginGetConnectionsWithConnectedMachineForFocusedMachineOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginGetConnectionsWithConnectedMachineForFocusedMachine"
@@ -492,7 +490,87 @@ func (client *MapsClient) getConnectionsWithConnectedMachineForFocusedMachineCre
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Content-Type"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, body); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// BeginGetDependencyViewForAllMachines - Get dependencies for all machines
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2025-07-01-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - mapName - Maps resource name
+//   - body - The content of the action request
+//   - options - MapsClientBeginGetDependencyViewForAllMachinesOptions contains the optional parameters for the MapsClient.BeginGetDependencyViewForAllMachines
+//     method.
+func (client *MapsClient) BeginGetDependencyViewForAllMachines(ctx context.Context, resourceGroupName string, mapName string, body GetDependencyViewForAllMachinesRequest, options *MapsClientBeginGetDependencyViewForAllMachinesOptions) (*runtime.Poller[MapsClientGetDependencyViewForAllMachinesResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.getDependencyViewForAllMachines(ctx, resourceGroupName, mapName, body, options)
+		if err != nil {
+			return nil, err
+		}
+		poller, err := runtime.NewPoller(resp, client.internal.Pipeline(), &runtime.NewPollerOptions[MapsClientGetDependencyViewForAllMachinesResponse]{
+			Tracer: client.internal.Tracer(),
+		})
+		return poller, err
+	} else {
+		return runtime.NewPollerFromResumeToken(options.ResumeToken, client.internal.Pipeline(), &runtime.NewPollerFromResumeTokenOptions[MapsClientGetDependencyViewForAllMachinesResponse]{
+			Tracer: client.internal.Tracer(),
+		})
+	}
+}
+
+// GetDependencyViewForAllMachines - Get dependencies for all machines
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2025-07-01-preview
+func (client *MapsClient) getDependencyViewForAllMachines(ctx context.Context, resourceGroupName string, mapName string, body GetDependencyViewForAllMachinesRequest, options *MapsClientBeginGetDependencyViewForAllMachinesOptions) (*http.Response, error) {
+	var err error
+	const operationName = "MapsClient.BeginGetDependencyViewForAllMachines"
+	ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, operationName)
+	ctx, endSpan := runtime.StartSpan(ctx, operationName, client.internal.Tracer(), nil)
+	defer func() { endSpan(err) }()
+	req, err := client.getDependencyViewForAllMachinesCreateRequest(ctx, resourceGroupName, mapName, body, options)
+	if err != nil {
+		return nil, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return nil, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusAccepted) {
+		err = runtime.NewResponseError(httpResp)
+		return nil, err
+	}
+	return httpResp, nil
+}
+
+// getDependencyViewForAllMachinesCreateRequest creates the GetDependencyViewForAllMachines request.
+func (client *MapsClient) getDependencyViewForAllMachinesCreateRequest(ctx context.Context, resourceGroupName string, mapName string, body GetDependencyViewForAllMachinesRequest, _ *MapsClientBeginGetDependencyViewForAllMachinesOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DependencyMap/maps/{mapName}/getDependencyViewForAllMachines"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if mapName == "" {
+		return nil, errors.New("parameter mapName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{mapName}", url.PathEscape(mapName))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -505,7 +583,7 @@ func (client *MapsClient) getConnectionsWithConnectedMachineForFocusedMachineCre
 // BeginGetDependencyViewForFocusedMachine - Get dependency map of single machine
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - body - The content of the action request
@@ -531,7 +609,7 @@ func (client *MapsClient) BeginGetDependencyViewForFocusedMachine(ctx context.Co
 // GetDependencyViewForFocusedMachine - Get dependency map of single machine
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) getDependencyViewForFocusedMachine(ctx context.Context, resourceGroupName string, mapName string, body GetDependencyViewForFocusedMachineRequest, options *MapsClientBeginGetDependencyViewForFocusedMachineOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginGetDependencyViewForFocusedMachine"
@@ -573,9 +651,8 @@ func (client *MapsClient) getDependencyViewForFocusedMachineCreateRequest(ctx co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
 		return nil, err
@@ -585,7 +662,7 @@ func (client *MapsClient) getDependencyViewForFocusedMachineCreateRequest(ctx co
 
 // NewListByResourceGroupPager - List MapsResource resources by resource group
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - MapsClientListByResourceGroupOptions contains the optional parameters for the MapsClient.NewListByResourceGroupPager
 //     method.
@@ -628,7 +705,7 @@ func (client *MapsClient) listByResourceGroupCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -645,7 +722,7 @@ func (client *MapsClient) listByResourceGroupHandleResponse(resp *http.Response)
 
 // NewListBySubscriptionPager - List MapsResource resources by subscription ID
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - options - MapsClientListBySubscriptionOptions contains the optional parameters for the MapsClient.NewListBySubscriptionPager
 //     method.
 func (client *MapsClient) NewListBySubscriptionPager(options *MapsClientListBySubscriptionOptions) *runtime.Pager[MapsClientListBySubscriptionResponse] {
@@ -683,7 +760,7 @@ func (client *MapsClient) listBySubscriptionCreateRequest(ctx context.Context, _
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -701,7 +778,7 @@ func (client *MapsClient) listBySubscriptionHandleResponse(resp *http.Response) 
 // BeginUpdate - Update a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - mapName - Maps resource name
 //   - properties - The resource properties to be updated.
@@ -726,7 +803,7 @@ func (client *MapsClient) BeginUpdate(ctx context.Context, resourceGroupName str
 // Update - Update a MapsResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-01-31-preview
+// Generated from API version 2025-07-01-preview
 func (client *MapsClient) update(ctx context.Context, resourceGroupName string, mapName string, properties MapsResourceTagsUpdate, options *MapsClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "MapsClient.BeginUpdate"
@@ -768,7 +845,7 @@ func (client *MapsClient) updateCreateRequest(ctx context.Context, resourceGroup
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-01-31-preview")
+	reqQP.Set("api-version", "2025-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}